refactor(DslTest): use containsExactly
